Move df building methods up into DataAssistantResult
